Support 'slow animation' setting for Transitioning API (#250)
We would like iOS simulator's "slow animation" mode to work with Transitions. Apparently, it only applies to UIView animations and not to lower level Core Animations. However, there is a way to detect that mode is turned ON and slow down animations manually by adjusting their duration. This solution is copied directly from RN core where it's been done for Animated.