[V2]: Avoid processing color when a processed value is provided (#888)
This change makes it possible for `useAnimatedStyle` to assign processed color (formatted into a number). This makes it possible for `processColor` to be used directly in `useAnimatedStyle`. Previously, we'd process the number which resulted in an undefined behavior. When color property already has a number assigned we assume it is processed and pass that directly as a prop to component.