Handle rerender behaviour (#1042)
## Description

This is supposed to provide a proper way to handle React Native's [Fast refresh](https://reactnative.dev/docs/fast-refresh) and also rerender of a component containing certain animations.

## Problem 🦯 

On component rerender all of the worklets which are included into that component are not rebuilt. This introduces following problems:
1. If someone changes worklet's body it will not be updated using React Native's [Fast refresh](https://reactnative.dev/docs/fast-refresh)
2. When worklet uses component's state and this state changes(causing rerender) that worklet will use outdated state value
<details>
<summary>ad. 2 example</summary>

``` JS
const Comp = () => {
  const [state, setState] = useState(0)
  const udv = useDerivedValue(() => {
    return state
  })
  return (
    <>
      <Button title="press" onPress={ () => {
        setState(state + 1)
      } } />
      <Text>Value(will not update): { udv.value }</Text>
    </>
  )
}
```
</details>

## Solutions

### Solution 1 👎 

As a first solution we came up with an idea of rebuilding everything on every hook call considering:
- `useSharedValue`
- `useAnimatedStyle`
- `useDerivedValue`

However after running some performance tests we noticed that `useSharedValue`'s performance has not changed that much but as for the other mentioned hooks they have slowed significantly. That led us to another solution.

### Solution 2 👍 

This is what is presented in this PR.
Shared value is reconstructed in `useSharedValue` when:
- it has not been created yet
- the hook is called with a different init value than in the previous render

`useAnimatedStyle` and `useDerivedValue` act different(but in the same manner as each other). They receive an array of dependencies like `useEffect` hook. The worklet inside is being reconstructed in following cases:
- it has not been created yet(this is first hook call)
- any provided dependencies changed(or there have not been any provided which is equal to listen to any change - for now `null` has to be passed, default is `[]`)
- worklet hash changed(which means worklet body changed)

### Changes

In `Hooks.js` three hooks has been modified as described above.
`FastRefreshTest.js` a testing screen has been added to check some of the most common use cases.