Improve caching. (#1236)
## WHY
Currently, when we want to use an object in two worklets then we have to covert the object twice. To avoid it this pr saves our jsi::HostObject as a hidden property (Object.defineProperty enumerable: false). Additionally, this pr introduces a different approach to representing FrozenObject and RemoteObject. JSC has problems with jsi::Objects that are created from HostObjects so instead of using them directly we expose frozen(not true with RemoteObject) jsi::Object and save jsi::HostObject as a hidden property.  

## HOW
During makeSherable operation add a hidden property to all objects converted to FrozenObject and WorkletFunction. 
`const char *HIDDEN_HOST_OBJECT_PROP = "__reanimatedHostObjectRef";`
Remove get/set from RemoteObject and return jsi::Object on UI side and frozenObj on JS side. Remove get from FrozenObject and use it only as a C++ representation. We only add a hidden property on JS side. When we convert jsi::Object on UI side in order to send it to RN we don't want to freeze it because it would be torture and it's a rare case.


## BREAKING CHANGE
From now on, all captured objects are frozen. It means if you want to pass a mutable object to the worklet then you have to copy it before.