Update ShareableValue.cpp (#1266)
The pr fixes some bugs introduced in #1236. 
We should only adapt the cache when object has already been converted.
We can't adapt caches when value is a  worklet because it's jest a function after conversion so we don't know if it's host function or not. Also fixed problem with jsThis creation which was caused by removing get method from frozenObject.