Simplify the iOS installation process and turn off turbo modules for other modules than Reanimated. (#1323)
This pr makes the installation of Reanimated on iOS similar to the mechanism we already used on Android. I've found a place where we can add Reanimated module to the global object (just like we do on Android). The changes let us install automatically on iOS. Another advantage of this approach is avoiding turbo versions of react-native libraries. It's important because some packages on npm are not ready yet for turbo versions. (e.g. react-navigation).

# HOW
We add the category of UIResponder class which is the parent class of `AppDelegate` which is used as a bridge delegate.
In this way, we don't need to know the name of `AppDelegate `class. As a result, we can add `jsExecutorFactoryForBridge` method to BridgeDelegate and install there reanimated module.  Additionally, we check if someone set flag `DONT_AUTOINTSTALL_REANIMATED`. If so we don't add category.