Add runOnJS logic. (#1272)
# WHY
When you call a function on UI thread you can't be sure if you call a worklet or a callback from JS thread. To make it more transparent we introduce `runOnJS` method which calls a callback asynchronously. If you call a callback without this method then an exception will be thrown.

## HOW
add a hidden property to every hostFunction named `callAsync` and call it in runOnJS.

resolves: https://github.com/software-mansion/react-native-reanimated/issues/1258